Version control (#17)
* feat: Showing tracing's logs

Default to warn level and progress until vvv -> trace

* feat: Adding username as argument to load data

* cfg: Adding tracing-subscriber to CLI crate

* Missing definition of argument username

* style:

* cfg: Updating CLI cargo.lock to include tracing-subscriber

* Split logging messages, one per variable

* refact: Function to load into DB expect the DB connection

Have clear intentions, and clear types.

* style:

* feat: Defining Ordinance data version with a global variable

* fix: Bad syntax for a comment within SQL statement

* Adding more log info

* fix: database must be mutable to add new data

* loading scrapped content within a transaction

The full load should be a single transaction: all or nothing; and
associated to one DB commit.

* Transmitting username from CLI

Need the username to fill the commit opperation.

* Adding some notes for the future

* Scrapper module must be public (for now)

That is required by the CLI

* Reverting back scrapper module to private

Let's try to minimize the public API as much as possible.

* Just a concept of async loader

* style:

* Recording some ideas

* refact: Moving scrapper.rs into a directory module

In preparation to break scrapper in smaller parts to make it easier to
follow.

* refact: Moving ScrapperConfig as a sub-module

* doc: scrapper::config top description

* fix: ScrapperConfig::from_json is required at the scrapper level

* refact: Moving ScrapperUsage into its own sub-module

Module scrapper was getting crowded and hard to read.

* test: Introducing the concept of samples, for ScrapperConfig

Isolate context while building samples for tests. The ScrappedOrdinance
is based on several components. Here we start to isolate the context
knowledge of each component in its own sub-modules. For instance,
crate::scrapper::config takes care to understand and have samples for
ScrapperConfig, while levels above that rely on support functions from
crate::scrapper::config::sample .

* test: Implementing samples for scrapper::usage

* cfg: Missing CLI dependency on duckDB

* feat: ScrapperConfig::init_db

Initialize database to support ScrapperConfig.

* feat: ScrapperConfig::write

Support inserting a ScrapperConfig into the DB.

* fix: ScrapperConfig::write with &Transaction

* log: Adding log trace for ScrapperConfig

* feat: ScrapperUsage::init_db

Initialize database for ScrapperUsage.

* refact: Renaming Usage to ScrapperUsage

* clean: Already moved those instructions into module usage

* refact, log: Using tracing explicitly

* feat: ScrappedOrdinance::init_db

* refact: Renamed to load_ordinance()

* log: Inform verbosity level as info

Up to level info, inform the verbosity level, otherwise, the defaul is
warning only and in that case don't bother notifying.

* feat: async ScrapperConfig::open()

A simplified initial implementation that doesn't take much advantage of
async for now, but instead returns a ScrapperConfig fully defined. This
is a proof of concept allowing the upper components to run all the way.
This will broken in multiple async components in the following steps.

* cfg: Missing features on tokio (fs + io)

* feat: Adding ScrapperConfig as an explicit component

* log: Notifing ScrapperUsage was written in the DB

* feat: async ScrapperUsage::open()

* feat: ScrapperUsage::write()

* refact: Renaming core usage table

* fix: Missing some 'use'

* log: Notifying DB initialized for ScrapperUsage

* feat: Including usage as an explicit component

* style:

* refact: Simplifying call to initialize DB

* style:

* load_ordinance can fail, thus return Result

* refact: Transfering db writing of each component

Instead of processing on the top level, now each component deals with
its context, thus, how and where to insert.

* fix: Get a commit_id for a new entry in bookkeeping

The regular execute ignores the return and instead shows the number of
rows changed, while what we want is the 'RETURNING id'.

* clean:

* fix: Connection must be mutable

* refact: ScrappedOrdinance::push

* feat: config and usage entries linked to a bookkeeping entry

* feat: Adding commit_id as an argument

Just a proof of concept. This will be implicit in a new
Commit/Transaction struct.

* doc: Saving some ideas and plans

* wip: Tests used to guide development

Once the design is estabilished, this should be cleaned.

* refact: Renaming table to bookkeeper

* test: Releasing test (wip)

* style:

* style: Unecessary use

For now, I'll use tracing's macro explicitly.