Update usage format (#65)
* refact: Removing total_time & total_time_seconds

The usage.json data model changed by removing the fields total_time &
total_time_seconds. This is the minimal change to make it work. The
current data structure in three levels might not make sense anymore but
let's keep as it is until we're sure we are not changing usage again.

* test: Improving test while reflecting the new structure

* refact: Adding an extra field to allow future flexibility

* style: