Filter by technology when exporting (#184)
* feat: Enum to define the technologies covered

The scraping process is currently focused on single technologies per
run. Although all the data is stored together in the database, it is
expected to be exported per technology. This new enum centralized a
consistent list of operating technologies, for the lib and the command
line.

* refact: CLI using Technology enum

* refact: Tech as &str

Let's capture technology as a str and pass that to core lib for now
while keeping `Technology` private for now.

The ideal would be an enum in the core library and use that in CLI as an
enum, so we have to update a single place to be used everywhere. But for
now, let's keep `Technology` private and postpone this.

* feat: impl Display for Technology

* clean: Don't need lifetime

* fix: Restricting export by technology

* style:

* style:

* cfg: Bump library version

Required since export is public and added new argument: technology.

* Passing format but not using it yet

* log: Adding more debug level information

* Replacing QuantitativeRecord by OrdinanceRecord

Struct used to handle quantitative + qualitative records.

* Adding bookkeeping to allow linking with technology

* feat: Let's have a specific format to satisfy reVX requirement

* doc: OrdinanceRecord

* feat: try_from(&str) for Technology

Cleaner process.